don't report an error if classes directory doesn't exist
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@6354 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ acqEngine/build.bat
@@ -1,4 +1,4 @@
-rmdir classes /s /q
+if exist classes rmdir classes /s /q
 mkdir classes
 xcopy /E /y src classes\
 "%JAVA_HOME%\bin\java" -cp ../../3rdpartypublic/classext/clojure.jar;../MMCoreJ_wrap/MMCoreJ.jar;../../3rdpartypublic/classext/ij.jar;../mmstudio/MMJ_.jar;./src -Dclojure.compile.path=classes clojure.lang.Compile org.micromanager.acq-engine
